Created Test Connection Database; Reorginized Solution; Created Database
