+++ Habraken-SLE/Classes/Designer/Designer.cs
