+++ Habraken-SLE/Classes/Label/Label.cs
