--- Habraken-SLE/Database.cs
@@ -1,43 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Habraken_SLE
-{
-    public class Database
-    {
-        public Database()
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public string connection
-        {
-            get => default(int);
-            set
-            {
-            }
-        }
-
-        public void TestConnection()
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public void OpenConnection()
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public void CloseConnection()
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public void Query()
-        {
-            throw new System.NotImplementedException();
-        }
-    }
-}
